refactor(monaco): simplify ata api
