--- packages/cdn/package.json
@@ -1,17 +1,0 @@
-{
-	"name": "@volar/cdn",
-	"version": "2.0.0-alpha.14",
-	"license": "MIT",
-	"files": [
-		"**/*.js",
-		"**/*.d.ts"
-	],
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/volarjs/volar.js.git",
-		"directory": "packages/cdn"
-	},
-	"dependencies": {
-		"@volar/language-service": "2.0.0-alpha.14"
-	}
-}
