+++ packages/language-service/lib/types.ts
@@ -1,17 +1,16 @@
 import type { LanguageContext } from '@volar/language-core';
 import type * as vscode from 'vscode-languageserver-protocol';
 import type { TextDocument } from 'vscode-languageserver-textdocument';
-import type { URI } from 'vscode-uri';
 import type { DocumentProvider } from './documents';
 
 export type * from 'vscode-languageserver-protocol';
 
 export interface ServiceEnvironment {
-	typescript: {
+	workspaceFolder: string;
+	typescript?: {
 		uriToFileName(uri: string): string;
 		fileNameToUri(fileName: string): string;
 	};
-	workspaceFolder: URI;
 	locale?: string;
 	clientCapabilities?: vscode.ClientCapabilities;
 	fs?: FileSystem;
